requirements.txt
fastapi
uvicorn
numpy
pandas
scikit-learn
librosa
praat-parselmouth
nltk
sentence-transformers
joblib
